Add position to the category-product-relation
Because there is the usecase, that a product should be assigned a
position within a category directly during import (to prevent it
appearing directly at the top of a well sorted category directly
after the import) and the vanilla import has a 'position' => 1
hardcoded this method was overridden.
The already existing processCategoryProductImport is not affected
but that one needs different datasets (like path and mostly strings
instead of category_ids) and another import run for all affected
products. I think in the end it is a matter of preference which one
you use but now you can at least choose and be no longer dependant
on a not-so-sensible-default like the hardcoded "1".